+++ src/snowflake/cli/api/project/schemas/entities/snowpark_entity.py
@@ -14,15 +14,18 @@
 
 from __future__ import annotations
 
-from typing import Dict, List, Literal, Optional, Union
+from typing import List, Literal, Optional, Union
 
 from pydantic import Field, field_validator
-from snowflake.cli.api.project.schemas.entities.common import EntityModelBase
+from snowflake.cli.api.project.schemas.entities.common import (
+    EntityModelBase,
+    ExternalAccessBaseModel,
+)
 from snowflake.cli.api.project.schemas.snowpark.argument import Argument
 from snowflake.cli.api.project.schemas.updatable_model import DiscriminatorField
 
 
-class SnowparkEntityModel(EntityModelBase):
+class SnowparkEntityModel(EntityModelBase, ExternalAccessBaseModel):
     handler: str = Field(
         title="Function’s or procedure’s implementation of the object inside source module",
         examples=["functions.hello_function"],
@@ -35,14 +38,6 @@ class SnowparkEntityModel(EntityModelBase):
     )
     runtime: Optional[Union[str, float]] = Field(
         title="Python version to use when executing ", default=None
-    )
-    external_access_integrations: Optional[List[str]] = Field(
-        title="Names of external access integrations needed for this procedure’s handler code to access external networks",
-        default=[],
-    )
-    secrets: Optional[Dict[str, str]] = Field(
-        title="Assigns the names of secrets to variables so that you can use the variables to reference the secrets",
-        default={},
     )
     imports: Optional[List[str]] = Field(
         title="Stage and path to previously uploaded files you want to import",