+++ RELEASE-NOTES.md
@@ -32,6 +32,7 @@
 * Added `snow spcs service execute-job` command, which supports creating and executing a job service in the current schema.
 * Added `snow app events` command to fetch logs and traces from local and customer app installations
 * Added support for project definition file defaults in templates
+* Added support for external access (api integrations and secrets) in Streamlit.
 
 ## Fixes and improvements
 * Fixed problem with whitespaces in `snow connection add` command